requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
SQLAlchemy==2.0.25
validators==0.22.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
